+++ libcnb-test/src/build.rs
@@ -8,8 +8,8 @@ use libcnb_package::dependency_graph::{get_dependencies, GetDependenciesError};
 use libcnb_package::output::create_packaged_buildpack_dir_resolver;
 use libcnb_package::{find_cargo_workspace_root_dir, CargoProfile, FindCargoWorkspaceRootError};
 use std::collections::BTreeMap;
-use std::fs;
 use std::path::{Path, PathBuf};
+use std::{fs, io};
 
 /// Packages the current crate as a buildpack into the provided directory.
 // TODO: Convert the `assert!` usages to an additional `PackageBuildpackError` variant instead:
@@ -23,11 +23,11 @@ pub(crate) fn package_crate_buildpack(
 ) -> Result<PathBuf, PackageBuildpackError> {
     let buildpack_toml = cargo_manifest_dir.join("buildpack.toml");
 
-    assert!(
-        buildpack_toml.exists(),
-        "Could not package directory as buildpack! No `buildpack.toml` file exists at {}",
-        cargo_manifest_dir.display()
-    );
+    if !buildpack_toml.exists() {
+        return Err(PackageBuildpackError::BuildpackDescriptorNotFound(
+            buildpack_toml,
+        ));
+    }
 
     let buildpack_descriptor: BuildpackDescriptor = read_toml_file(buildpack_toml)
         .map_err(PackageBuildpackError::CannotReadBuildpackDescriptor)?;
@@ -53,7 +53,7 @@ pub(crate) fn package_buildpack(
 ) -> Result<PathBuf, PackageBuildpackError> {
     let cargo_build_env = match cross_compile_assistance(target_triple.as_ref()) {
         CrossCompileAssistance::HelpText(help_text) => {
-            return Err(PackageBuildpackError::CrossCompileConfigurationError(
+            return Err(PackageBuildpackError::CrossCompileToolchainNotFound(
                 help_text,
             ));
         }
@@ -75,28 +75,21 @@ pub(crate) fn package_buildpack(
 
     let root_node = buildpack_dependency_graph
         .node_weights()
-        .find(|node| node.buildpack_id == buildpack_id.clone());
+        .find(|node| &node.buildpack_id == buildpack_id)
+        .ok_or_else(|| {
+            PackageBuildpackError::BuildpackIdNotFound(buildpack_id.clone(), workspace_root_path)
+        })?;
 
-    assert!(
-        root_node.is_some(),
-        "Could not package directory as buildpack! No buildpack with id `{buildpack_id}` exists in the workspace at {}",
-        workspace_root_path.display()
-    );
-
-    let build_order = get_dependencies(
-        &buildpack_dependency_graph,
-        &[root_node.expect("The root node should exist")],
-    )
-    .map_err(PackageBuildpackError::GetDependencies)?;
+    let build_order = get_dependencies(&buildpack_dependency_graph, &[root_node])
+        .map_err(PackageBuildpackError::GetDependencies)?;
 
     let mut packaged_buildpack_dirs = BTreeMap::new();
     for node in &build_order {
         let buildpack_destination_dir = buildpack_dir_resolver(&node.buildpack_id);
 
-        // TODO: Convert the `unwrap()` to an additional `PackageBuildpackError` variant instead:
-        // https://github.com/heroku/libcnb.rs/issues/710
-        #[allow(clippy::unwrap_used)]
-        fs::create_dir_all(&buildpack_destination_dir).unwrap();
+        fs::create_dir_all(&buildpack_destination_dir).map_err(|error| {
+            PackageBuildpackError::CannotCreateDirectory(buildpack_destination_dir.clone(), error)
+        })?;
 
         libcnb_package::package::package_buildpack(
             &node.path,
@@ -114,12 +107,24 @@ pub(crate) fn package_buildpack(
     Ok(buildpack_dir_resolver(buildpack_id))
 }
 
-#[derive(Debug)]
+#[derive(thiserror::Error, Debug)]
 pub(crate) enum PackageBuildpackError {
+    #[error("Couldn't find a buildpack.toml file at {0}")]
+    BuildpackDescriptorNotFound(PathBuf),
+    #[error("Couldn't find a buildpack with ID '{0}' in the workspace at {1}")]
+    BuildpackIdNotFound(BuildpackId, PathBuf),
+    #[error("I/O error creating directory {0}: {1}")]
+    CannotCreateDirectory(PathBuf, io::Error),
+    #[error("Couldn't read buildpack.toml: {0}")]
     CannotReadBuildpackDescriptor(TomlFileError),
+    #[error("Couldn't calculate buildpack dependency graph: {0}")]
     BuildBuildpackDependencyGraph(BuildBuildpackDependencyGraphError),
-    CrossCompileConfigurationError(String),
+    #[error("Couldn't find cross-compilation toolchain.\n\n{0}")]
+    CrossCompileToolchainNotFound(String),
+    #[error("Couldn't find Cargo workspace root: {0}")]
     FindCargoWorkspaceRoot(FindCargoWorkspaceRootError),
+    #[error("Couldn't get buildpack dependencies: {0}")]
     GetDependencies(GetDependenciesError<BuildpackId>),
+    #[error(transparent)]
     PackageBuildpack(libcnb_package::package::PackageBuildpackError),
 }